upload additional file

+++ database/migrations/2021_03_17_160449_create_dichvus_table.php
@@ -16,8 +16,8 @@ class CreateDichvusTable extends Migration
         Schema::create('dichvus', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name_service');
-            $table->decimal('price');
-            $table->decimal('sale_price');
+            $table->double('price');
+            $table->double('sale_price');
             $table->string('description');
             $table->string('image');
             $table->integer('id_cate');
